Split up common types and hooks
