+++ labextension/src/widgets/hooks/useEnableByDefaultEffect.ts
@@ -6,6 +6,11 @@ interface IUseEnableByDefaultEffectParams {
   setIsEnabled: (value: boolean | ((prev: boolean) => boolean)) => void;
 }
 
+/**
+ * Hook that reacts to the "Enable Kale by default" JupyterLab setting.
+ * When the setting transitions from off to on, it turns on the Kale
+ * toggle for the current notebook without requiring a notebook switch.
+ */
 export function useEnableByDefaultEffect({
   enableKaleByDefault,
   isEnabled,
